added case for reader where it will not exit on error, only print the error message.
This is because of the opensuse bug making /etc/issue a symlinked file
to root only readable /run/issue, a bug I can't fix, and which is unlikely
to be corrected. Since I have no way of knowing when someone in a distro is
going to break all history and tradition and practices, it's probably best to
just make reader print error and not exit.