Updated distro detections, the various derived distros are morphing their methods of ID, several are now using os-release with the ID_LIKE field.
antergos just changed, for example, and now uses os-release for antergos. This is
kind of random sadly, and handling dynamically morphing methods without messing up
the ID is not easy, sad to say. But it does appear the distros are starting to move
towards os-release becoming more usable, but it can only be used after testing its
contents to confirm if it's old or new style, sadly. the difficulty the free software
ecosystem has in the most trivial task of creating a simple text file with a simple
format that would handle all derived and system base data is truly astounding!