big update on disk vendors, added new logic where testing via actual vendor sub routine is done. This saves time per run through now.
Optimized a bit of code in hard coded loops over arrays of arrays,
now not using assigning reference to array variable, but using reference
directly. This may or may not save memory and execution speed on large
loops, results suggest maybe a tiny savings though I'd have to test it more
robustly to really know.

Not using this method where the data in the arrays is not known, and not
using on hashes for the time being.