Many small updates, enhancements, bug fixes!!! We've been saving them up!! Here they are!! Don't wait!!
Thanks mr. mazda for many issue finds, and suggestions.

--------------------------------------------------------------------------------
KNOWN ISSUES BUT CAN'T OR WON'T BE FIXED:

1. Due to unfixable rpm slowdowns, removed package counts for default output for
rpm based systems. We were seeing delays of up to 30+ seconds just to list the
rpm package count, which is absurd, even after the rpm optimizations inxi
already runs. To allow rpm users to get excluded by default for rpm package list
counts, added --pkg flag plus a  short message telling them to use that flag to
get the installed package count if they want it.

Changes like this are very unfortunate, but in 2021 for a package manager at
times to require over 30 seconds to generate a trivial installed package list is
just not acceptable. One of the reasons this release was delayed was this was
not an easy decision to make, it's very rare support for a feature is removed
for specific tools due to how badly the tools may perform. Note that whatever
higher level tool is used, like dnf, zypp, it's still the same speed, they all
appear to use the same core engine.

Basically this decision was forced since either inxi looks really bad and slow,
when it's not, or the actual cause was removed from default outputs.

--------------------------------------------------------------------------------
BUGS:

1. Small bug in nfs blacklist for disk used led to nfs used being added, which
leads to silly used percents. This is corrected.

2. If ram vendor ID failed, inxi would delete the part number. Oops. This was
related to the Mushkin failures.

3. Close to a bug, though not one internally, but to users would appear as one:
ZFS does not act as expected, zpool list did not in fact return the pool size,
which I had always assumed to be the case, but in a very strange decision, does
return something very close to the pool size for mirrors, but NOT for z1 or z2
pools, then it returns the total size of the drives that make up the pool. To
call this strange behavior would be an understatement. The fix was to modify the
logic to use zfs list instead to get the size data. This also makes the drive
total report far more accurate, since it lists usable space now for ZFS as was
always intended. The cause of this was simply that I'd always had access to zfs
mirrors, not z1 or z2 arrays.

--------------------------------------------------------------------------------
FIXES:

1. OpenSuse and maybe others use kdm3 for Trinity, not kdm, so dm was failing.

2. Going along with fix 1, made kde version detection more robust so may catch
more fringe / corner cases for kde desktops. These were mainly added to correct
Trinity desktop version detections.

3. Mushkin ram vendor ID was failing, that is or should be corrected.

4. Added in /dev/disk/by-id handlers for zpool components, there are several
variants, wwn-, pci-, scsi-, ata-, but they all map to the real /dev drive IDs.
Failure to unmap these led to failing to match components and get size info
etc for zfs.

5. See DOCUMENTATION: 2, language changes for weather feature abuse.

--------------------------------------------------------------------------------
ENHANCEMENTS:

1. Going along with the rpm issues, added dnf.conf support to yum/dnf repo
types. Not sure how that one was missed, but there it is. This should tighten
repo reports for dnf/yum/zypp types.

2. Added LeftWM. LeftWM confirmed working. Added unverifed detections for:
penrose, 2bwm, 5dwm, catwm, mcwm, monsterwm, snapwm, uwm, wingo, wmfs, wmfs2.

3. Added xfwm as a compositor type, that had bee left out, somewhat on purpose,
since xfwm can run in compositing or non compositing mode. But should show
since many users use compositing mode now.

4. Added OpenMediaVault distro ID and systembase handlers.

5. Going along with zfs bug fix 3, using zfs list data for free, size,
allocated. Trying to understand how zfs developers actually thought about this
is nearly impossible so just used what seems to correspond to reality most.
Also shows raw values for zfs data in RAID along with regular ones to make
clear which is which value.

6. Added more CPU architecture ID matches for AMD Zen and a variety of Intel.
Both vendors finally released some new CPUs and the data became available,
which doesn't always happen quickly.

7. A bunch of new disk vendors and vendor IDs added. Never stops, like the
sands of time, like the ocean waves, like the scuttling crabs scrounding around
in the seaweed in the foam where the outgoing wave left its mark...

--------------------------------------------------------------------------------
DOCUMENTATION:

1. Added leftwm keybindinigs to inxi-data.txt desktop/wm section. Updated more
wm in that section as well, and list more info on wms for future reference etc.
Also reorganized and more more readable wm section.

2. Help/Man now make more clear that automated requests or excessive use of the
inxi weather feature are not under any circumstance permitted. There had been
some ambiguity and lack of clarity about what abuse is, now it should be more
clear.

--------------------------------------------------------------------------------
CODE CHANGES:

1. Refactored uptime parser logic, the code and regex was just getting too
messy and difficult to work with and debug, now it works similar to how the
revised BSD parsers run, the regex are pulled apart and made more granular so
a small syntax change ideally won't break the detections as easily.

2. Cleaned up sub cpu_arch() and made all the arch values line up nicely, over
time I notice that almost invariably stuff done to save lines of code makes
code harder to read as the feature expands, so it's generally worth just
unravelling it so it all stacks and is easy to scan/read. Also removed extra
white space in parens, which is something I'm leaning more towards but it's
not worth fixing all at once so it's just done where it's noticed.

That's using:
if ( /test/ ){
rather than:
if (/test){

I believe using more white space helped with Perl comprehension in the
intermediate stages, but is not required anymore and just looks like extra
whitespace now.